Revert "feat: new mobile wallet popover P1 (#10395)"
This reverts commit b536cdd3b8dde2f1fd9309894952f33e64febd9c.
