+++ .env
@@ -51,7 +51,6 @@ VITE_FEATURE_NEW_LIMIT_FLOW=true
 VITE_FEATURE_THORCHAIN_SWAPPER_ACK=false
 VITE_FEATURE_ACTION_CENTER=true
 VITE_FEATURE_QUICK_BUY=false
-VITE_FEATURE_NEW_WALLET_MANAGER=false
 VITE_FEATURE_CHATWOOT=true
 
 # absolute URL prefix
