+++ .env.development
@@ -7,7 +7,6 @@
 # feature flags
 VITE_FEATURE_THORCHAIN_TCY_ACTIVITY=true
 VITE_FEATURE_QUICK_BUY=true
-VITE_FEATURE_NEW_WALLET_MANAGER=true
 
 # mixpanel
 VITE_MIXPANEL_TOKEN=a867ce40912a6b7d01d088cf62b0e1ff
