+++ .env.production
@@ -3,7 +3,6 @@ VITE_FEATURE_MIXPANEL=true
 
 VITE_FEATURE_THORCHAIN_TCY_ACTIVITY=false
 VITE_FEATURE_QUICK_BUY=false
-VITE_FEATURE_NEW_WALLET_MANAGER=false
 
 # mixpanel
 VITE_MIXPANEL_TOKEN=9d304465fc72224aead9e027e7c24356
