+++ src/components/Layout/Header/Header.tsx
@@ -11,7 +11,6 @@ import { GlobalSeachButton } from './GlobalSearch/GlobalSearchButton'
 import { ChainMenu } from './NavBar/ChainMenu'
 import { MobileNavBar } from './NavBar/MobileNavBar'
 import { UserMenu } from './NavBar/UserMenu'
-import { WalletManagerPopover } from './NavBar/WalletManagerPopover'
 import { TxWindow } from './TxWindow/TxWindow'
 
 import { useDiscoverAccounts } from '@/context/AppProvider/hooks/useDiscoverAccounts'
@@ -54,7 +53,6 @@ export const Header = memo(() => {
 
   const isWalletConnectToDappsV2Enabled = useFeatureFlag('WalletConnectToDappsV2')
   const isActionCenterEnabled = useFeatureFlag('ActionCenter')
-  const isNewWalletManagerEnabled = useFeatureFlag('NewWalletManager')
   const { degradedChainIds } = useDiscoverAccounts()
 
   /**
@@ -121,7 +119,7 @@ export const Header = memo(() => {
               {isConnected && isActionCenterEnabled && <ActionCenter />}
               {isLargerThanMd && (
                 <Box display={displayProp2}>
-                  {isNewWalletManagerEnabled ? <WalletManagerPopover /> : <UserMenu />}
+                  <UserMenu />
                 </Box>
               )}
             </Flex>
