+++ src/components/Layout/Header/NavBar/UserMenu.tsx
@@ -1,14 +1,31 @@
-import { ChevronDownIcon } from '@chakra-ui/icons'
+import { ChevronDownIcon, WarningTwoIcon } from '@chakra-ui/icons'
 import type { ComponentWithAs, IconProps } from '@chakra-ui/react'
-import { Box, ButtonGroup, Menu, MenuGroup, MenuItem, MenuList } from '@chakra-ui/react'
-import { memo, useCallback } from 'react'
+import {
+  Box,
+  Button,
+  ButtonGroup,
+  Flex,
+  HStack,
+  Menu,
+  MenuButton,
+  MenuGroup,
+  MenuItem,
+  MenuList,
+  useColorModeValue,
+} from '@chakra-ui/react'
+import type { FC } from 'react'
+import { memo, useCallback, useEffect, useMemo, useState } from 'react'
+import { FaWallet } from 'react-icons/fa'
 import { useTranslate } from 'react-polyglot'
 import { MemoryRouter } from 'react-router-dom'
-
-import { WalletButton } from './WalletButton'
+import { getAddress } from 'viem'
+import { useEnsName } from 'wagmi'
 
 import { WalletConnectedRoutes } from '@/components/Layout/Header/NavBar/hooks/useMenuRoutes'
 import { WalletConnectedMenu } from '@/components/Layout/Header/NavBar/WalletConnectedMenu'
+import { WalletImage } from '@/components/Layout/Header/NavBar/WalletImage'
+import { MiddleEllipsis } from '@/components/MiddleEllipsis/MiddleEllipsis'
+import { RawText, Text } from '@/components/Text'
 import { WalletActions } from '@/context/WalletProvider/actions'
 import type { InitialState } from '@/context/WalletProvider/WalletProvider'
 import { useWallet } from '@/hooks/useWallet/useWallet'
@@ -21,6 +38,7 @@ export const entries = [WalletConnectedRoutes.Connected]
 
 const maxWidthProp = { base: 'full', md: 'xs' }
 const minWidthProp = { base: 0, md: 'xs' }
+const widthProp = { base: '100%', lg: 'auto' }
 
 const NoWallet = ({ onClick }: { onClick: () => void }) => {
   const translate = useTranslate()
@@ -59,6 +77,103 @@ export const WalletConnected = (props: WalletConnectedProps) => {
   )
 }
 
+type WalletButtonProps = {
+  isConnected: boolean
+  isLoadingLocalWallet: boolean
+  onConnect: () => void
+} & Pick<InitialState, 'walletInfo'>
+
+const WalletButton: FC<WalletButtonProps> = ({
+  isConnected,
+  walletInfo,
+  onConnect,
+  isLoadingLocalWallet,
+}) => {
+  const [walletLabel, setWalletLabel] = useState('')
+  const [shouldShorten, setShouldShorten] = useState(true)
+  const bgColor = useColorModeValue('gray.200', 'gray.800')
+
+  const { data: ensName } = useEnsName({
+    address: walletInfo?.meta?.address ? getAddress(walletInfo.meta.address) : undefined,
+  })
+
+  const maybeRdns = useAppSelector(selectWalletRdns)
+
+  const mipdProviders = useMipdProviders()
+  const maybeMipdProvider = useMemo(
+    () => mipdProviders.find(provider => provider.info.rdns === maybeRdns),
+    [mipdProviders, maybeRdns],
+  )
+
+  useEffect(() => {
+    setWalletLabel('')
+    setShouldShorten(true)
+    if (!walletInfo || !walletInfo.meta) return setWalletLabel('')
+    // Wallet has a native label, we don't care about ENS name here
+    if (!walletInfo?.meta?.address && walletInfo.meta.label) {
+      setShouldShorten(false)
+      return setWalletLabel(walletInfo.meta.label)
+    }
+
+    // ENS is registered for address and is successfully fetched. Set ENS name as label
+    if (ensName) {
+      setShouldShorten(false)
+      return setWalletLabel(ensName)
+    }
+
+    // No label or ENS name, set regular wallet address as label
+    return setWalletLabel(walletInfo?.meta?.address ?? '')
+  }, [ensName, walletInfo])
+
+  const rightIcon = useMemo(() => <ChevronDownIcon />, [])
+  const leftIcon = useMemo(
+    () => (
+      <HStack>
+        {!isConnected && <WarningTwoIcon ml={2} w={3} h={3} color='yellow.500' />}
+        <WalletImage walletInfo={maybeMipdProvider?.info || walletInfo} />
+      </HStack>
+    ),
+    [isConnected, maybeMipdProvider, walletInfo],
+  )
+  const connectIcon = useMemo(() => <FaWallet />, [])
+
+  const handleMenuClick = useCallback(() => {
+    vibrate('heavy')
+  }, [])
+
+  return Boolean(walletInfo?.deviceId) || isLoadingLocalWallet ? (
+    <MenuButton
+      as={Button}
+      width={widthProp}
+      justifyContent='flex-start'
+      rightIcon={rightIcon}
+      leftIcon={leftIcon}
+      onClick={handleMenuClick}
+    >
+      <Flex>
+        {walletLabel ? (
+          <MiddleEllipsis
+            rounded='lg'
+            fontSize='sm'
+            p='1'
+            pl='2'
+            pr='2'
+            shouldShorten={shouldShorten}
+            bgColor={bgColor}
+            value={walletLabel}
+          />
+        ) : (
+          <RawText>{walletInfo?.name}</RawText>
+        )}
+      </Flex>
+    </MenuButton>
+  ) : (
+    <Button onClick={onConnect} leftIcon={connectIcon}>
+      <Text translation='common.connectWallet' />
+    </Button>
+  )
+}
+
 export const UserMenu: React.FC<{ onClick?: () => void }> = memo(({ onClick }) => {
   const { state, dispatch, disconnect } = useWallet()
   const { isConnected, walletInfo, connectedType, isLocked, isLoadingLocalWallet } = state
@@ -79,7 +194,6 @@ export const UserMenu: React.FC<{ onClick?: () => void }> = memo(({ onClick }) =
       <Box>
         <Menu autoSelect={false}>
           <WalletButton
-            isMenuContext
             onConnect={handleConnect}
             walletInfo={walletInfo}
             isConnected={isConnected && !isLocked}
