+++ src/test/mocks/store.ts
@@ -161,7 +161,6 @@ export const mockStore: ReduxState = {
       LazyTxHistory: false,
       RfoxFoxEcosystemPage: false,
       QuickBuy: false,
-      NewWalletManager: false,
     },
     quickBuyAmounts: [10, 50, 100],
     quoteDisplayOption: QuoteDisplayOption.Basic,
